#41 #72 introduce AbstractIntegrationTest
